Turn on CONFIG_NO_HZ_IDLE
This saves a lot of time for large, idle VMs.  Just running the timer
interrupt takes a lot of cycles.  The numbers vary based on your host
cpu, but I saw about 1.5% of a host cpu required to drive a 250 Hz tick
for a *single* vcpu.  The time spent was about 3:1 host-to-guest, almost
all kernel time.

If you're curious, I did something like this:

	perf stat -e cycles:u -e cycles:k -e cycles:h -e cycles:H \
	          -e cycles:G -p $VMM_PID -- sleep 10

Signed-off-by: Barret Rhoden <brho@google.com>