Switch from tinycore to u-root
u-root does just about whatever we need.  It's small, has sshd, etc.
It's ability to add extra files is better (via Pox).

We lose the PACKAGES stuff, but it's not a great loss.  I attempted to
keep it around - you could still install the openssh tcz into tc_root
- but without the base distro of tinycore, you're missing out on shared
libraries, and it's a losing battle.

Of course, maybe we will need this, and I'll be cursing u-root's sshd
and missing openssh...

Signed-off-by: Barret Rhoden <brho@google.com>